+++ pkg/components/ebpf/common/sql_detect_mysql.go
@@ -5,7 +5,11 @@ package ebpfcommon
 
 import (
 	"encoding/binary"
+	"log/slog"
 	"strings"
+
+	"go.opentelemetry.io/obi/pkg/app/request"
+	"go.opentelemetry.io/obi/pkg/components/sqlprune"
 )
 
 type mysqlPreparedStatementsKey struct {
@@ -78,4 +82,82 @@ func mysqlPreparedStatements(b []byte) (string, string, string) {
 	sql := text
 
 	return op, table, sql
+}
+
+func handleMySQL(parseCtx *EBPFParseContext, event *TCPRequestInfo, requestBuffer, responseBuffer []byte) (request.Span, error) {
+	var (
+		op, table, stmt string
+		span            request.Span
+	)
+
+	if len(requestBuffer) < sqlprune.MySQLHdrSize+1 {
+		slog.Warn("MySQL request too short")
+		return span, errFallback
+	}
+	if len(responseBuffer) < sqlprune.MySQLHdrSize+1 {
+		slog.Warn("MySQL response too short")
+		return span, errFallback
+	}
+
+	sqlCommand := sqlprune.SQLParseCommandID(request.DBMySQL, requestBuffer)
+	if sqlCommand == "" {
+		return span, errIgnore
+	}
+
+	sqlError := sqlprune.SQLParseError(responseBuffer)
+
+	switch sqlCommand {
+	case "STMT_PREPARE":
+		if sqlError != nil {
+			slog.Debug("MySQL PREPARE command errored, ignoring", "error", sqlError)
+			return span, errIgnore
+		}
+
+		// On the PREPARE command, the statement ID is the first 4 bytes after the header and command ID
+		// in the response buffer.
+		stmtID := sqlprune.SQLParseStatementID(request.DBMySQL, responseBuffer)
+		if stmtID == 0 {
+			slog.Warn("MySQL PREPARE command with invalid statement ID")
+			return span, errFallback
+		}
+
+		_, _, stmt = detectSQL(string(requestBuffer[sqlprune.MySQLHdrSize+1:]))
+		parseCtx.mysqlPreparedStatements.Add(mysqlPreparedStatementsKey{
+			connInfo: event.ConnInfo,
+			stmtID:   stmtID,
+		}, stmt)
+
+		return span, errIgnore
+	case "STMT_EXECUTE":
+		// On the EXECUTE command, the statement ID is the first 4 bytes after the header and command ID
+		// in the request buffer.
+		stmtID := sqlprune.SQLParseStatementID(request.DBMySQL, requestBuffer)
+		if stmtID == 0 {
+			slog.Warn("MySQL EXECUTE command with invalid statement ID")
+			return span, errFallback
+		}
+
+		var found bool
+		stmt, found = parseCtx.mysqlPreparedStatements.Get(mysqlPreparedStatementsKey{
+			connInfo: event.ConnInfo,
+			stmtID:   stmtID,
+		})
+		if !found {
+			slog.Debug("MySQL EXECUTE command with unknown statement ID", "stmtID", stmtID)
+			return span, errFallback
+		}
+		op, table = sqlprune.SQLParseOperationAndTable(stmt)
+	case "QUERY":
+		op, table, stmt = detectSQL(string(requestBuffer[sqlprune.MySQLHdrSize+1:]))
+	default:
+		slog.Warn("MySQL command ID unhandled", "commandID", requestBuffer[sqlprune.MySQLHdrSize])
+		return span, errFallback
+	}
+
+	if !validSQL(op, table, request.DBMySQL) {
+		slog.Warn("MySQL operation and/or table are invalid", "stmt", stmt)
+		return span, errFallback
+	}
+
+	return TCPToSQLToSpan(event, op, table, stmt, request.DBMySQL, sqlCommand, sqlError), nil
 }
