+++ pkg/components/ebpf/common/tcp_large_buffer_test.go
@@ -86,18 +86,7 @@ func TestTCPLargeBuffers(t *testing.T) {
 	// The buffer should now be firstBuf + appendBuf
 	verifyLargeBuffer(firstEvent.Tp.TraceId, firstEvent.Tp.SpanId, firstEvent.Direction, firstBuf+appendBuf)
 
-	// Test append to non-existing buffer (should just add a new one)
-	newTraceID := [16]uint8{'3'}
-	newSpanID := [8]uint8{'4'}
-	appendEvent.Tp.TraceId = newTraceID
-	appendEvent.Tp.SpanId = newSpanID
-	_, _, err = appendTCPLargeBuffer(pctx, toRingbufRecord(t, appendEvent, appendBuf))
-	require.NoError(t, err)
-	verifyLargeBuffer(newTraceID, newSpanID, firstEvent.Direction, appendBuf)
-
 	// Test multiple appends
-	appendEvent.Tp.TraceId = firstEvent.Tp.TraceId
-	appendEvent.Tp.SpanId = firstEvent.Tp.SpanId
 	// Re-init buffer
 	firstEvent.Len = uint32(len(firstBuf))
 	_, _, err = appendTCPLargeBuffer(pctx, toRingbufRecord(t, firstEvent, firstBuf))
