+++ pkg/obi/config.go
@@ -78,6 +78,7 @@ var DefaultConfig = Config{
 			MySQL: 0,
 		},
 		MySQLPreparedStatementsCacheSize: 1024,
+		MongoRequestsCacheSize:           1024,
 	},
 	NameResolver: &transform.NameResolverConfig{
 		Sources:  []string{"k8s"},
