+++ test/integration/test_utils.go
@@ -241,11 +241,6 @@ func waitForTestComponentsRoute(t *testing.T, url, route string) {
 	}, test.Interval(time.Second))
 }
 
-//nolint:unused
-func waitForSQLTestComponentsPostgres(t *testing.T, url, subpath string) {
-	waitForSQLTestComponentsWithDB(t, url, subpath, "postgresql")
-}
-
 func waitForSQLTestComponentsMySQL(t *testing.T, url, subpath string) {
 	waitForSQLTestComponentsWithDB(t, url, subpath, "mysql")
 }
